refactor: split Config Debug impl into config_debug.rs; drop stale allowlist
The repo monolith gate counts effective lines (no comments/blanks/test mods),
under which config.rs (219), config_impls.rs (498), and config_snapshot.rs (486)
were all already ≤500 — the earlier allowlist entries were unnecessary
(CodeRabbit's 800/510/505 were raw line counts). Remove those three entries.

config_impls.rs sat at 498 effective and grows with every Config field, so move
the large `impl fmt::Debug for Config` (secret redaction) into a sibling
config_debug.rs (config_impls.rs 498→262, config_debug.rs 239). Also add the
previously-missing warc_output/automation_script fields to the Debug output.
