+++ src/core/config/types.rs
@@ -1,4 +1,5 @@
 mod config;
+mod config_debug;
 mod config_impls;
 mod enums;
 pub mod overrides;
